Implemented the gather neighbor step from Sean’s Connected Components algorithm, after hours of trying. Z-VALUES IN gl_Position HAVE TO BE BETWEEN 0 AND 1!!!

+++ OGLTest/IndexBuffer.h
@@ -19,7 +19,7 @@ public:
     
     void Bind();
     
-    void SetData(const List<GLushort> &indices);
+    void SetData(const List<GLuint> &indices);
     
     GLsizei Count() const;
     
@@ -37,10 +37,10 @@ inline void IndexBuffer::Bind()
     BindTo(GL_ELEMENT_ARRAY_BUFFER);
 }
 
-inline void IndexBuffer::SetData(const List<GLushort> &indices)
+inline void IndexBuffer::SetData(const List<GLuint> &indices)
 {
     count = (GLsizei)indices.size();
-    OGLBuffer::SetData(GL_ELEMENT_ARRAY_BUFFER, count * sizeof(GLushort), indices.data());
+    OGLBuffer::SetData(GL_ELEMENT_ARRAY_BUFFER, count * sizeof(GLuint), indices.data());
 }
 
 inline GLsizei IndexBuffer::Count() const
